allow capture intervals to be set in fractions of a second

+++ core/options.cpp
@@ -338,7 +338,7 @@ Options::Options()
 			"Set the brightness of the blue channel")
 		("monochrome", value<bool>(&v_->monochrome)->default_value(false)->implicit_value(true),
 			"fixes awb, sets dng metadata to monochrome")
-		("capture-interval", value<unsigned int>(&v_->capture_interval)->default_value(0),
+		("capture-interval", value<float>(&v_->capture_interval)->default_value(0.0f),
 			"Sets the interval between captures in seconds")
 		("force-jpeg", value<bool>(&v_->force_jpeg)->default_value(false)->implicit_value(true),
 			"Force the use of the JPEG encoder")
